refactor tests to use shared qt stubs
