+++ tests/test_account_categories.py
@@ -5,209 +5,12 @@ import pandas as pd
 import unittest
 from decimal import Decimal
 
+from tests.qt_stubs import patch_qt_modules
+
 from src.utils.account_categories import CategoryCalculator
 
 FIXTURES = os.path.join(os.path.dirname(__file__), "fixtures")
 
-
-def patch_qt_modules():
-    """Provide minimal PyQt6 stubs so AccountCategoryDialog can be imported."""
-
-    class Signal:
-        def __init__(self):
-            self._slots = []
-
-        def connect(self, slot):
-            self._slots.append(slot)
-
-        def emit(self, *args, **kwargs):
-            for s in list(self._slots):
-                s(*args, **kwargs)
-
-    class Qt:
-        class ItemFlag:
-            ItemIsUserCheckable = 1
-
-        class CheckState:
-            Unchecked = 0
-            Checked = 2
-
-    class QListWidgetItem:
-        def __init__(self, text=""):
-            self._text = text
-            self._flags = 0
-            self._check_state = Qt.CheckState.Unchecked
-
-        def text(self):
-            return self._text
-
-        def setFlags(self, flags):
-            self._flags = flags
-
-        def flags(self):
-            return self._flags
-
-        def setCheckState(self, state):
-            self._check_state = state
-
-        def checkState(self):
-            return self._check_state
-
-    class QListWidget:
-        def __init__(self):
-            self.items = []
-            self.currentItemChanged = Signal()
-            self.itemChanged = Signal()
-            self._current = None
-
-        def addItems(self, items):
-            for text in items:
-                self.addItem(QListWidgetItem(text))
-
-        def addItem(self, item):
-            if isinstance(item, str):
-                item = QListWidgetItem(item)
-            self.items.append(item)
-
-        def item(self, index):
-            return self.items[index]
-
-        def count(self):
-            return len(self.items)
-
-        def clear(self):
-            self.items.clear()
-
-        def blockSignals(self, flag):
-            pass
-
-        def setCurrentItem(self, item):
-            prev = self._current
-            self._current = item
-            self.currentItemChanged.emit(item, prev)
-
-        def currentItem(self):
-            return self._current
-
-        def setCurrentRow(self, row):
-            if 0 <= row < len(self.items):
-                self.setCurrentItem(self.items[row])
-            else:
-                self.setCurrentItem(None)
-
-        def row(self, item):
-            try:
-                return self.items.index(item)
-            except ValueError:
-                return -1
-
-        def takeItem(self, row):
-            if 0 <= row < len(self.items):
-                return self.items.pop(row)
-
-    class QWidget:
-        def __init__(self, *args, **kwargs):
-            pass
-
-        def addWidget(self, *args, **kwargs):
-            pass
-
-        def addLayout(self, *args, **kwargs):
-            pass
-
-    class QDialog(QWidget):
-        def accept(self):
-            pass
-
-        def reject(self):
-            pass
-
-    class QVBoxLayout(QWidget):
-        pass
-
-    class QHBoxLayout(QWidget):
-        pass
-
-    class QPushButton(QWidget):
-        def __init__(self, *args, **kwargs):
-            super().__init__()
-            self.clicked = Signal()
-
-    class QLineEdit(QWidget):
-        def __init__(self, *args, **kwargs):
-            super().__init__()
-            self._text = ""
-            self.textChanged = Signal()
-
-        def text(self):
-            return self._text
-
-        def setText(self, text):
-            self._text = text
-            self.textChanged.emit(text)
-
-        def clear(self):
-            self.setText("")
-
-    class QLabel(QWidget):
-        pass
-
-    class QTabWidget(QWidget):
-        def addTab(self, *args, **kwargs):
-            pass
-
-    class QDialogButtonBox(QWidget):
-        class StandardButton:
-            Ok = 1
-            Cancel = 2
-
-        def __init__(self, *args, **kwargs):
-            super().__init__()
-            self.accepted = Signal()
-            self.rejected = Signal()
-
-    class QMessageBox:
-        class StandardButton:
-            Save = 1
-            Discard = 2
-            Cancel = 3
-
-        next_result = StandardButton.Save
-
-        @staticmethod
-        def question(*args, **kwargs):
-            return QMessageBox.next_result
-
-    class QInputDialog:
-        @staticmethod
-        def getText(*args, **kwargs):
-            return "", False
-
-    widgets = types.ModuleType("PyQt6.QtWidgets")
-    for name, obj in {
-        "QDialog": QDialog,
-        "QVBoxLayout": QVBoxLayout,
-        "QHBoxLayout": QHBoxLayout,
-        "QListWidget": QListWidget,
-        "QListWidgetItem": QListWidgetItem,
-        "QPushButton": QPushButton,
-        "QLineEdit": QLineEdit,
-        "QInputDialog": QInputDialog,
-        "QMessageBox": QMessageBox,
-        "QLabel": QLabel,
-        "QTabWidget": QTabWidget,
-        "QWidget": QWidget,
-        "QDialogButtonBox": QDialogButtonBox,
-    }.items():
-        setattr(widgets, name, obj)
-
-    core = types.ModuleType("PyQt6.QtCore")
-    core.Qt = Qt
-
-    sys.modules.setdefault("PyQt6", types.ModuleType("PyQt6"))
-    sys.modules["PyQt6.QtWidgets"] = widgets
-    sys.modules["PyQt6.QtCore"] = core
-    sys.modules.setdefault("PyQt6.QtGui", types.ModuleType("PyQt6.QtGui"))
 
 
 class DummyConfig:
